Ajuste da estrutura do projeto
Ajuste no local das classes e estrutura geral do projeto.
