+++ tre/src/mg/puc/modelos/ListaUrna.java
@@ -1,4 +1,6 @@
-package mg.puc.urna;
+package mg.puc.modelos;
+
+import mg.puc.Urna;
 
 /***
  * Representa uma lista de urnas.
@@ -52,5 +54,45 @@ public class ListaUrna {
 	 */
 	public void reiniciarInteracao() {
 		this.itemAtual = null;
+	}
+	
+	/***
+	 * Representa um item na lista de urnas. A classe é declarada como privada, dentro de ListaUrna, por
+	 * somente ser utilizada como apoio para a lista.
+	 * @author rafaella-matos
+	 */
+	private class ItemListaUrna {
+		private ItemListaUrna proximo = null;
+		private Urna urna = null;
+		
+		/***
+		 * Obtém o próximo registro da lista.
+		 */
+		public ItemListaUrna getProximo() {
+			return this.proximo;
+		}
+		
+		/***
+		 * Configura a referência ao item seguinte a este.
+		 * @param item A referência ao item posterior a este.
+		 */
+		public void setProximo(ItemListaUrna item) {
+			this.proximo = item;
+		}
+		
+		/***
+		 * Obtém a urna associada a esse item da lista.
+		 */
+		public Urna getUrna() {
+			return this.urna;
+		}
+		
+		/***
+		 * Initializa uma nova instância de Urna.
+		 * @param urna A urna para inicialização.
+		 */
+		public ItemListaUrna(Urna urna) {
+			this.urna = urna;
+		}	
 	}
 }
