--- urna/src/mg/puc/urna/Main.java
@@ -1,8 +1,0 @@
-package mg.puc.urna;
-
-public class Main {
-
-  public static void main(String[] args) {
-    // write your code here
-  }
-}
